# Make Collection.set_paths even faster