--- a/lib/matplotlib/collections.py
+++ b/lib/matplotlib/collections.py
@@ -1070,6 +1070,10 @@ class PolyCollection(_CollectionWithSizes):
         self.set_verts(verts, closed)
         self.stale = True
 
+    # This function creates a lot of Path object, which is pretty much
+    # guaranteed to trigger the GC multiple times. None of them will be garbage
+    # just yet, so all those runs are completely unnecessary.
+    @cbook._without_gc
     def set_verts(self, verts, closed=True):
         """
         Set the vertices of the polygons.
@@ -1094,15 +1098,19 @@ class PolyCollection(_CollectionWithSizes):
             return
 
         # Fast path for arrays
-        if isinstance(verts, np.ndarray):
-            verts_pad = np.concatenate((verts, verts[:, :1]), axis=1)
+        if isinstance(verts, np.ndarray) and len(verts):
+            verts_pad = (np.concatenate((verts, verts[:, :1]), axis=1)
+                           .astype(mpath.Path.verts_type))
             # Creating the codes once is much faster than having Path do it
             # separately each time by passing closed=True.
-            codes = np.empty(verts_pad.shape[1], dtype=mpath.Path.code_type)
-            codes[:] = mpath.Path.LINETO
-            codes[0] = mpath.Path.MOVETO
-            codes[-1] = mpath.Path.CLOSEPOLY
-            self._paths = [mpath.Path(xy, codes) for xy in verts_pad]
+            example_path = mpath.Path(verts_pad[0], closed=True)
+            # Looking up the values once speeds up the iteration a bit
+            _make_path = mpath.Path._fast_from_codes_and_verts
+            codes = example_path.codes
+            self._paths = [_make_path(xy, codes,
+                                      internals_from=example_path,
+                                      unmask_verts=False)
+                           for xy in verts_pad]
             return
 
         self._paths = []