--- a/lib/matplotlib/path.py
+++ b/lib/matplotlib/path.py
@@ -74,8 +74,12 @@ class Path:
         made up front in the constructor that will not change when the
         data changes.
     """
+    __slots__ = ('_vertices', '_codes', '_readonly',
+                 '_should_simplify', '_simplify_threshold',
+                 '_interpolation_steps', '__weakref__')
 
     code_type = np.uint8
+    verts_type = float
 
     # Path codes
     STOP = code_type(0)         # 1 vertex
@@ -160,7 +164,8 @@ class Path:
             self._readonly = False
 
     @classmethod
-    def _fast_from_codes_and_verts(cls, verts, codes, internals_from=None):
+    def _fast_from_codes_and_verts(cls, verts, codes, internals_from=None,
+                                   unmask_verts=True):
         """
         Creates a Path instance without the expense of calling the constructor.
 
@@ -173,9 +178,17 @@ class Path:
             ``should_simplify``, ``simplify_threshold``, and
             ``interpolation_steps`` will be copied.  Note that ``readonly`` is
             never copied, and always set to ``False`` by this constructor.
+        unmask_verts : bool
+            If False, vertices should be an unmasked NumPy array with dtype
+            set to Path.verts_type.
         """
         pth = cls.__new__(cls)
-        pth._vertices = _to_unmasked_float_array(verts)
+        # NOTE: _to_unmasked_float_array has non-trivial overhead in tight
+        #       loops, so we allow skipping it if the caller wants to
+        if unmask_verts:
+            pth._vertices = _to_unmasked_float_array(verts)
+        else:
+            pth._vertices = verts
         pth._codes = codes
         pth._readonly = False
         if internals_from is not None:
@@ -268,16 +281,6 @@ class Path:
         `True` if the `Path` is read-only.
         """
         return self._readonly
-
-    def __copy__(self):
-        """
-        Returns a shallow copy of the `Path`, which will share the
-        vertices and codes with the source `Path`.
-        """
-        import copy
-        return copy.copy(self)
-
-    copy = __copy__
 
     def __deepcopy__(self, memo=None):
         """