--- a/lib/matplotlib/tests/test_transforms.py
+++ b/lib/matplotlib/tests/test_transforms.py
@@ -621,28 +621,6 @@ def test_invalid_arguments():
         t.transform([[1, 2, 3]])
 
 
-def test_transformed_path():
-    points = [(0, 0), (1, 0), (1, 1), (0, 1)]
-    path = Path(points, closed=True)
-
-    trans = mtransforms.Affine2D()
-    trans_path = mtransforms.TransformedPath(path, trans)
-    assert_allclose(trans_path.get_fully_transformed_path().vertices, points)
-
-    # Changing the transform should change the result.
-    r2 = 1 / np.sqrt(2)
-    trans.rotate(np.pi / 4)
-    assert_allclose(trans_path.get_fully_transformed_path().vertices,
-                    [(0, 0), (r2, r2), (0, 2 * r2), (-r2, r2)],
-                    atol=1e-15)
-
-    # Changing the path does not change the result (it's cached).
-    path.points = [(0, 0)] * 4
-    assert_allclose(trans_path.get_fully_transformed_path().vertices,
-                    [(0, 0), (r2, r2), (0, 2 * r2), (-r2, r2)],
-                    atol=1e-15)
-
-
 def test_transformed_patch_path():
     trans = mtransforms.Affine2D()
     patch = mpatches.Wedge((0, 0), 1, 45, 135, transform=trans)