## PR Summary

As in the title, this PR optimizes `Collection.set_paths` even further. With all PRs I've submitted before this one applied, the rendering for my example (400x400 surface plot on a 2012 MBP) takes ~2s on average. With this patch, the time drops to ~0.7s. The breakdown of this 0.7s is:

* ~0.4s in `Poly3DCollection.do_3d_projection`
  * Including ~0.2s in `Collection.set_paths`
* ~0.3s in `Collection.draw` (mostly spend in the C code for the Agg backend in this case)

## PR Checklist

- [x] Has Pytest style unit tests
- [x] Code is [Flake 8](http://flake8.pycqa.org/en/latest/) compliant
- [x] New features are documented, with examples if plot related
- [x] Documentation is sphinx and numpydoc compliant
- [x] Added an entry to doc/users/next_whats_new/ if major new feature (follow instructions in README.rst there)
- [x] Documented in doc/api/api_changes.rst if API changed in a backward-incompatible way