@apaszke 2020-03-25
@QuLogic This should have addressed all your comments.