@jklymak 2020-09-10
ping @apaszke still interested in this optimization?