@tacaswell 2023-05-24
Coming back to this, I am 👍🏻 on the `__slots__`, but 👎🏻 on adding a decorator to disable gc.  I'm mixed on inlining `_make_path.