@scottshambaugh 2025-01-07
I've picked this up in https://github.com/matplotlib/matplotlib/pull/29398, building off this PR (but without the garbage collection change). Thank you @apaszke for the initial work here!